requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
ConcurrentLogHandler==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.1
Django==2.1.3
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-configurations==2.1
django-suit==0.2.26
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.15.0
ephem==3.7.6.0
eventlet==0.24.1
Fabric3==1.14.post1
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
monotonic==1.5
olefile==0.46
paramiko==2.4.2
Pillow==5.3.0
psycopg2==2.7.6.1
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.6.4
PyNaCl==1.3.0
python-crontab==2.3.5
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
